Flytter vilkårperiode-kode fra VilkårService/Controller til Vilkårperiode-spesifike klasser - Nye endepunkter for vilkårperiode
