+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/stønadsvilkår/VilkårService.kt
@@ -2,12 +2,10 @@ package no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår
 
 import no.nav.tilleggsstonader.kontrakter.felles.Stønadstype
 import no.nav.tilleggsstonader.sak.behandling.BehandlingService
-import no.nav.tilleggsstonader.sak.behandling.BehandlingUtil.validerBehandlingIdErLik
 import no.nav.tilleggsstonader.sak.behandling.barn.BarnService
 import no.nav.tilleggsstonader.sak.behandling.barn.BehandlingBarn
 import no.nav.tilleggsstonader.sak.fagsak.FagsakService
 import no.nav.tilleggsstonader.sak.infrastruktur.database.Sporbar
-import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrThrow
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.Feil
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
 import no.nav.tilleggsstonader.sak.opplysninger.søknad.SøknadService
@@ -20,19 +18,6 @@ import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.tilDto
 import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.regler.HovedregelMetadata
 import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.regler.evalutation.OppdaterVilkår
 import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.regler.evalutation.OppdaterVilkår.opprettNyeVilkår
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.EvalueringVilkårperiode.evaulerVilkårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.KildeVilkårsperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.MålgruppeType
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.ResultatVilkårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeRepository
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeType
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.OpprettVilkårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.SlettVikårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.VilkårperiodeDto
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.Vilkårperioder
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.tilDto
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -43,7 +28,6 @@ class VilkårService(
     private val behandlingService: BehandlingService,
     private val søknadService: SøknadService,
     private val vilkårRepository: VilkårRepository,
-    private val vilkårperiodeRepository: VilkårperiodeRepository,
     private val barnService: BarnService,
     private val vilkårGrunnlagService: VilkårGrunnlagService,
     // private val grunnlagsdataService: GrunnlagsdataService,
@@ -74,59 +58,6 @@ class VilkårService(
             }*/
         }
         return hentEllerOpprettVilkårsvurdering(behandlingId)
-    }
-
-    fun hentVilkårperioder(behandlingId: UUID): Vilkårperioder {
-        val vilkårsperioder = vilkårperiodeRepository.findByBehandlingId(behandlingId)
-
-        return Vilkårperioder(
-            målgrupper = finnPerioder<MålgruppeType>(vilkårsperioder),
-            aktiviteter = finnPerioder<AktivitetType>(vilkårsperioder),
-        )
-    }
-
-    private inline fun <reified T : VilkårperiodeType> finnPerioder(
-        vilkårsperioder: List<Vilkårperiode>,
-    ) = vilkårsperioder.filter { it.type is T }.map(Vilkårperiode::tilDto)
-
-    @Transactional
-    fun opprettVilkårperiode(behandlingId: UUID, opprettVilkårperiode: OpprettVilkårperiode): VilkårperiodeDto {
-        feilHvis(behandlingErLåstForVidereRedigering(behandlingId)) {
-            "Kan ikke opprette vilkår når behandling er låst for videre redigering"
-        }
-
-        val resultatEvaluering = evaulerVilkårperiode(opprettVilkårperiode.delvilkår)
-        val vilkårperiode = vilkårperiodeRepository.insert(
-            Vilkårperiode(
-                behandlingId = behandlingId,
-                fom = opprettVilkårperiode.fom,
-                tom = opprettVilkårperiode.tom,
-                type = opprettVilkårperiode.type,
-                delvilkår = resultatEvaluering.delvilkår,
-                begrunnelse = opprettVilkårperiode.begrunnelse,
-                resultat = resultatEvaluering.resultat,
-                kilde = KildeVilkårsperiode.MANUELL,
-            ),
-        )
-
-        return vilkårperiode.tilDto()
-    }
-
-    fun slettVilkårperiode(behandlingId: UUID, id: UUID, slettVikårperiode: SlettVikårperiode): Vilkårperiode {
-        val vilkårperiode = vilkårperiodeRepository.findByIdOrThrow(id)
-
-        validerBehandlingIdErLik(behandlingId, vilkårperiode.behandlingId)
-
-        feilHvis(behandlingErLåstForVidereRedigering(vilkårperiode.behandlingId)) {
-            "Kan ikke slette vilkårperiode når behandling er låst for videre redigering"
-        }
-
-        return vilkårperiodeRepository.update(
-            vilkårperiode.copy(
-                resultat = ResultatVilkårperiode.SLETTET,
-                slettetKommentar = slettVikårperiode.kommentar,
-            ),
-        )
     }
 
     fun hentVilkårsett(behandlingId: UUID): List<VilkårDto> {
