+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/VilkårperiodeController.kt
@@ -1,0 +1,58 @@
+package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode
+
+import no.nav.security.token.support.core.api.ProtectedWithClaims
+import no.nav.tilleggsstonader.sak.tilgang.AuditLoggerEvent
+import no.nav.tilleggsstonader.sak.tilgang.TilgangService
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.OpprettVilkårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.SlettVikårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.VilkårperiodeDto
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.Vilkårperioder
+import org.springframework.validation.annotation.Validated
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+import java.util.UUID
+
+@RestController
+@RequestMapping(path = ["/api/vilkarperiode"])
+@ProtectedWithClaims(issuer = "azuread")
+@Validated
+class VilkårperiodeController(
+    private val tilgangService: TilgangService,
+    private val vilkårperiodeService: VilkårperiodeService,
+) {
+
+    @GetMapping("behandling/{behandlingId}")
+    fun hentMålgrupper(@PathVariable behandlingId: UUID): Vilkårperioder {
+        tilgangService.validerTilgangTilBehandling(behandlingId, AuditLoggerEvent.ACCESS)
+
+        return vilkårperiodeService.hentVilkårperioder(behandlingId)
+    }
+
+    @PostMapping("behandling/{behandlingId}")
+    fun opprettVilkårMedPeriode(
+        @PathVariable behandlingId: UUID,
+        @RequestBody opprettVilkårperiode: OpprettVilkårperiode,
+    ): VilkårperiodeDto {
+        tilgangService.validerTilgangTilBehandling(behandlingId, AuditLoggerEvent.UPDATE)
+        tilgangService.validerHarSaksbehandlerrolle()
+
+        return vilkårperiodeService.opprettVilkårperiode(behandlingId, opprettVilkårperiode)
+    }
+
+    @DeleteMapping("{id}")
+    fun slettPeriode(
+        @PathVariable("id") id: UUID,
+        @RequestBody slettVikårperiode: SlettVikårperiode,
+    ): Vilkårperiode {
+        tilgangService.validerTilgangTilBehandling(slettVikårperiode.behandlingId, AuditLoggerEvent.UPDATE)
+        tilgangService.validerHarSaksbehandlerrolle()
+
+        return vilkårperiodeService.slettVilkårperiode(id, slettVikårperiode)
+    }
+}
