+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/VilkårperiodeService.kt
@@ -1,0 +1,84 @@
+package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode
+
+import no.nav.tilleggsstonader.sak.behandling.BehandlingService
+import no.nav.tilleggsstonader.sak.behandling.BehandlingUtil.validerBehandlingIdErLik
+import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrThrow
+import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.KildeVilkårsperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.MålgruppeType
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.ResultatVilkårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeRepository
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeType
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.OpprettVilkårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.SlettVikårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.VilkårperiodeDto
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.Vilkårperioder
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.tilDto
+import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
+import java.util.UUID
+
+@Service
+class VilkårperiodeService(
+    private val behandlingService: BehandlingService,
+    private val vilkårperiodeRepository: VilkårperiodeRepository,
+) {
+
+    fun hentVilkårperioder(behandlingId: UUID): Vilkårperioder {
+        val vilkårsperioder = vilkårperiodeRepository.findByBehandlingId(behandlingId)
+
+        return Vilkårperioder(
+            målgrupper = finnPerioder<MålgruppeType>(vilkårsperioder),
+            aktiviteter = finnPerioder<AktivitetType>(vilkårsperioder),
+        )
+    }
+
+    private inline fun <reified T : VilkårperiodeType> finnPerioder(
+        vilkårsperioder: List<Vilkårperiode>,
+    ) = vilkårsperioder.filter { it.type is T }.map(Vilkårperiode::tilDto)
+
+    @Transactional
+    fun opprettVilkårperiode(behandlingId: UUID, opprettVilkårperiode: OpprettVilkårperiode): VilkårperiodeDto {
+        feilHvis(behandlingErLåstForVidereRedigering(behandlingId)) {
+            "Kan ikke opprette vilkår når behandling er låst for videre redigering"
+        }
+
+        val resultatEvaluering = EvalueringVilkårperiode.evaulerVilkårperiode(opprettVilkårperiode.delvilkår)
+        val vilkårperiode = vilkårperiodeRepository.insert(
+            Vilkårperiode(
+                behandlingId = behandlingId,
+                fom = opprettVilkårperiode.fom,
+                tom = opprettVilkårperiode.tom,
+                type = opprettVilkårperiode.type,
+                delvilkår = resultatEvaluering.delvilkår,
+                begrunnelse = opprettVilkårperiode.begrunnelse,
+                resultat = resultatEvaluering.resultat,
+                kilde = KildeVilkårsperiode.MANUELL,
+            ),
+        )
+
+        return vilkårperiode.tilDto()
+    }
+
+    fun slettVilkårperiode(id: UUID, slettVikårperiode: SlettVikårperiode): Vilkårperiode {
+        val vilkårperiode = vilkårperiodeRepository.findByIdOrThrow(id)
+
+        validerBehandlingIdErLik(slettVikårperiode.behandlingId, vilkårperiode.behandlingId)
+
+        feilHvis(behandlingErLåstForVidereRedigering(vilkårperiode.behandlingId)) {
+            "Kan ikke slette vilkårperiode når behandling er låst for videre redigering"
+        }
+
+        return vilkårperiodeRepository.update(
+            vilkårperiode.copy(
+                resultat = ResultatVilkårperiode.SLETTET,
+                slettetKommentar = slettVikårperiode.kommentar,
+            ),
+        )
+    }
+
+    private fun behandlingErLåstForVidereRedigering(behandlingId: UUID) =
+        behandlingService.hentBehandling(behandlingId).status.behandlingErLåstForVidereRedigering()
+}
