+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/dto/VilkårperiodeDto.kt
@@ -95,6 +95,7 @@ data class DelvilkårAktivitetDto(
 ) : DelvilkårVilkårperiodeDto()
 
 data class SlettVikårperiode(
+    val behandlingId: UUID,
     val kommentar: String,
 )
