+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/VilkårperiodeServiceTest.kt
@@ -1,0 +1,101 @@
+package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode
+
+import no.nav.tilleggsstonader.sak.IntegrationTest
+import no.nav.tilleggsstonader.sak.behandling.domain.BehandlingRepository
+import no.nav.tilleggsstonader.sak.behandling.domain.BehandlingStatus
+import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrThrow
+import no.nav.tilleggsstonader.sak.infrastruktur.sikkerhet.SikkerhetContext
+import no.nav.tilleggsstonader.sak.util.BrukerContextUtil
+import no.nav.tilleggsstonader.sak.util.behandling
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.KildeVilkårsperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.ResultatVilkårperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeDomainUtil
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeRepository
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.SlettVikårperiode
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+
+class VilkårperiodeServiceTest : IntegrationTest() {
+
+    @Autowired
+    lateinit var behandlingRepository: BehandlingRepository
+
+    @Autowired
+    lateinit var vilkårperiodeService: VilkårperiodeService
+
+    @Autowired
+    lateinit var vilkårperiodeRepository: VilkårperiodeRepository
+
+    @Nested
+    inner class SlettVilkårperiode {
+
+        @Test
+        fun `skal ikke kunne slette kommentar hvis behandlingen ikke er under behandling`() {
+            val behandling =
+                testoppsettService.opprettBehandlingMedFagsak(behandling(status = BehandlingStatus.FERDIGSTILT))
+            val målgruppe = VilkårperiodeDomainUtil.målgruppe(
+                behandlingId = behandling.id,
+                kilde = KildeVilkårsperiode.MANUELL,
+            )
+            val periode = vilkårperiodeRepository.insert(målgruppe)
+
+            Assertions.assertThatThrownBy {
+                vilkårperiodeService.slettVilkårperiode(periode.id, SlettVikårperiode(behandling.id, "kommentar"))
+            }.hasMessageContaining("Kan ikke slette vilkårperiode når behandling er låst for videre redigering")
+        }
+
+        @Test
+        fun `skal ikke kunne slette kommentar hvis man mangler kommentar`() {
+            val behandling =
+                testoppsettService.opprettBehandlingMedFagsak(behandling())
+            val målgruppe = VilkårperiodeDomainUtil.målgruppe(
+                behandlingId = behandling.id,
+                kilde = KildeVilkårsperiode.MANUELL,
+            )
+            val periode = vilkårperiodeRepository.insert(målgruppe)
+
+            Assertions.assertThatThrownBy {
+                vilkårperiodeService.slettVilkårperiode(periode.id, SlettVikårperiode(behandling.id, "    "))
+            }.hasMessageContaining("Mangler kommentar")
+        }
+
+        @Test
+        fun `skal ikke kunne slette kommentar hvis kilden er system`() {
+            val behandling =
+                testoppsettService.opprettBehandlingMedFagsak(behandling())
+            val målgruppe = VilkårperiodeDomainUtil.målgruppe(
+                behandlingId = behandling.id,
+                kilde = KildeVilkårsperiode.SYSTEM,
+            )
+            val periode = vilkårperiodeRepository.insert(målgruppe)
+
+            Assertions.assertThatThrownBy {
+                vilkårperiodeService.slettVilkårperiode(periode.id, SlettVikårperiode(behandling.id, "kommentar"))
+            }.hasMessageContaining("Kan ikke slette når kilde=")
+        }
+
+        @Test
+        fun `skal kunne slette kommentar som er manuellt opprettet`() {
+            val saksbehandler = "saksbehandlerX"
+            val behandling =
+                testoppsettService.opprettBehandlingMedFagsak(behandling())
+            val målgruppe = VilkårperiodeDomainUtil.målgruppe(
+                behandlingId = behandling.id,
+                kilde = KildeVilkårsperiode.MANUELL,
+            )
+            val periode = vilkårperiodeRepository.insert(målgruppe)
+
+            Assertions.assertThat(periode.sporbar.endret.endretAv).isEqualTo(SikkerhetContext.SYSTEM_FORKORTELSE)
+
+            BrukerContextUtil.testWithBrukerContext(saksbehandler) {
+                vilkårperiodeService.slettVilkårperiode(periode.id, SlettVikårperiode(behandling.id, "kommentar"))
+            }
+
+            val oppdatertPeriode = vilkårperiodeRepository.findByIdOrThrow(periode.id)
+            Assertions.assertThat(oppdatertPeriode.resultat).isEqualTo(ResultatVilkårperiode.SLETTET)
+            Assertions.assertThat(oppdatertPeriode.sporbar.endret.endretAv).isEqualTo(saksbehandler)
+        }
+    }
+}
